src/api/app/requirements2.txt
#PyICU==2.3.1
polyglot==16.07.04
#pycld2==0.41
six==1.12.0
regex==2021.8.3

# Transitive Dependencies

aiohttp==3.8.0
async-timeout==4.0.1
idna-ssl==1.1.0
multidict==5.1.0
yarl==1.6.3
click==7.1.2
fastapi==0.65.2
futures==3.1.1
h11==0.9.0
#httptools==0.1.1
JSON-log-formatter==0.3.0
Morfessor==2.0.6
numpy==1.22.0
orjson==3.5.3
prometheus-client==0.8.0
PyYAML==5.4.1
pydantic==1.8.2
starlette==0.14.2
starlette-exporter==0.4.1
symspellpy==6.7.0
typing-extensions==3.7.4.3
#uvicorn==0.14.0
#uvloop==0.14.0
websockets==10.0
wcwidth==0.2.5
autowrapt==1.0
oneagent-sdk==1.4.0.20210127.165413
wrapt==1.13.1



#######################################



--extra-index-url https://europe-python.pkg.dev/${PROJECT_NAME}/python-packages/simple/

thg-gcp-utils==1.1.0

cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
Deprecated==1.2.13
google-api-core==2.10.2
google-auth==2.9.1
google-cloud-core==2.3.2
google-cloud-monitoring==2.14.2
google-cloud-storage==2.7.0
google-cloud-trace==1.7.3
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
grpcio-tools==1.54.2
idna==3.4
importlib-metadata==6.6.0
jaraco.classes==3.2.3
keyring==23.13.1
keyrings.google-artifactregistry-auth==1.1.2
more-itertools==9.1.0
opentelemetry-api==1.16.0
opentelemetry-distro==0.37b0
opentelemetry-exporter-gcp-monitoring==1.4.0a0
opentelemetry-exporter-gcp-trace==1.4.0
opentelemetry-instrumentation==0.37b0
opentelemetry-sdk==1.16.0
opentelemetry-semantic-conventions==0.37b0
pluggy==1.0.0
proto-plus==1.22.2
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.7
PyYAML==6.0
requests==2.30.0
rsa==4.9
six==1.16.0
thg-gcp-utils==1.1.0
typing_extensions==4.5.0
urllib3==1.26.15
wrapt==1.15.0
zipp==3.15.0
